# Speed up read dimension fields in TS