--- a/x-pack/plugin/esql/compute/src/test/java/org/elasticsearch/compute/lucene/LuceneSourceOperatorTests.java
+++ b/x-pack/plugin/esql/compute/src/test/java/org/elasticsearch/compute/lucene/LuceneSourceOperatorTests.java
@@ -324,5 +324,10 @@ public class LuceneSourceOperatorTests extends AnyOperatorTestCase {
         public String shardIdentifier() {
             return "test";
         }
+
+        @Override
+        public MappedFieldType fieldType(String name) {
+            throw new UnsupportedOperationException();
+        }
     }
 }