When reading dimension fields in the TS command, we can skip reading values while the `tsid` remains unchanged to improve performance. I benchmarked this change with the following query:

```
POST /_query
{
    "query": "TS metrics-hostmetricsreceiver.otel-default | WHERE @timestamp >= \"2025-05-08T18:00:08.001Z\" | STATS cpu = avg(rate(`metrics.process.cpu.time`)) BY host.name, BUCKET(@timestamp, 5 minute)"
}
```

The total query time was reduced from 51ms to 39ms, with processing time in the time-series source operator reduced from 26ms to 17ms.